Add files via upload
The main object of this project is to help people who want to buy a house so they can know the price range and plan their finances in advance. It may also help those who are in the real estate business. Here I have done explanatory data analysis(EDA) and used various machine learning models to see which gives a better prediction.